Added v4. Moved files.
I added the v4 of the tray. This edits the space on the side to have more "extra parts" space. I also added text to make organization easier.

I moved the v2 files to a v2 folder for organization.

I added text to the thelio-io badge holder. Also, put four on one print for the respective modles.