Added a print, P1S gcode
E30 print was added. I added the original gcode for the pre print checks, and the modified gcode that skips most of the pre print checks. with the faststart.gcode, the P1S will not bed level, check vibration, or whatever else was there. Now it just homes, wipes, and prints :)