removed flush, initial extruder temp
I removed the HB fan and filament ooze protection. The extruder is over the pooper, and the fan is turned on shortly after this.

Removed the flush, wipe should take care of this, just double check the extruder is not clogged.

Removed another temperature set. It lowered the temp for the exturder which just makes it heat up again. the wipe should take care of clogging and what not. Double check after wipe